==> notebooks/2_data_analysis.ipynb <==
## **Анализ данных**

import numpy as np
import pandas as pd

### Загрузка данных

cases = pd.read_csv('../data/cases_labeled.csv', sep=';')
print(cases.columns.to_list())
cases.head()

marks = pd.read_csv('../data/marks.csv', sep=';')
print(marks.columns.to_list())
marks.head()

diaries = pd.read_csv('../data/diaries.csv', sep=';')
diaries['Затрачено часов'] = diaries['Затрачено часов'].str.replace(',', '.').astype(float)
print(diaries.columns.to_list())
diaries.head()

print('Всего выполненных поручений каждого типа:')
cases.loc[cases['Выполнено'] == 'да']['Тип поручения'].value_counts()

note_types = diaries['Запись в дневнике'].unique()
print('Все виды записей в дневнике:')
print(*note_types, sep=', ')
print()

class_to_type = {0:'Чудовище в пещере',
                 1:'Спасение путников',
                 2:'Поиск пропажи',
                 3:'монстры/разбойники'
                 }

for note in note_types:
    ids = diaries.loc[diaries['Запись в дневнике'] == note]['Номер поручения']
    texts = cases.loc[cases['Номер поручения'].isin(ids)]['Тип поручения'].value_counts().to_dict()
    print(f'"{note}" встречается в {len(texts)} типах поручений:')
    for key in texts:
        print(f'В заказе на "{class_to_type[key]}"({key}) встречается {texts[key]} раз')
    print()

print('=' * 50 + '\n')

role_types = diaries['Роль'].unique()
print('Все виды ролей:')
print(*role_types, sep=', ')
print()

class_to_role = {i:[] for i in range(4)}

for role in role_types:
    ids = diaries.loc[diaries['Роль'] == role]['Номер поручения']
    texts = cases.loc[cases['Номер поручения'].isin(ids)]['Тип поручения'].value_counts().to_dict()
    print(f'"{role}" встречается в {len(texts)} типах поручений:')
    for key in texts:
        class_to_role[key].append(role)
        print(f'В заказе на "{class_to_type[key]}"({key}) встречается {texts[key]} раз')
    print()

print('=' * 50 + '\n')
print(diaries[['Запись в дневнике', 'Роль']].groupby('Запись в дневнике').value_counts())

class_to_role

Средние оценки героев по всем поручениям:

heroes = marks['Герой'].unique()
print(heroes)

marks[['Герой','Оценка за качество','Оценка по срокам','Оценка за вежливость']].groupby('Герой').mean()

Количество записей в дневких для каждой роли героя

note_cnt = diaries[['Герой', 'Роль', 'Запись в дневнике']].groupby('Герой').value_counts()
for hero in heroes:
    print(hero)
    print(note_cnt[hero])
    print()

### Формирование датасета со статистикой оценок/времени/суммы по героям

diaries_edited = pd.merge(diaries, marks, on=['Номер поручения', 'Герой'], how='left')
diaries_edited = pd.merge(diaries_edited, cases[['Номер поручения', 'Тип поручения', 'Затрачено дней', 'Сумма вознаграждения']], on=['Номер поручения'], how='left')

stats = diaries_edited[['Герой', 'Тип поручения', 'Роль', 'Оценка за качество', 'Оценка по срокам', 'Оценка за вежливость', 'Затрачено часов', 'Затрачено дней', 'Сумма вознаграждения']].groupby(['Герой', 'Тип поручения', 'Роль']).mean()
stats['Количество поручений'] = diaries_edited[['Герой', 'Тип поручения', 'Роль']].groupby(['Герой', 'Тип поручения', 'Роль']).value_counts()
stats

stats.to_csv('../data/stats.csv')
