==> notebooks/3_create_team.ipynb <==
import pandas as pd
import numpy as np
from scipy.optimize import linprog
import joblib
from datetime import datetime

data = pd.read_csv('../data/stats.csv')
cases = pd.read_csv('../data/cases_labeled.csv', sep=';')
data

roles_for_tasks = {
    0: ['лекарь', 'лучник', 'мечник', 'боевой маг'],
    1: ['следопыт'],
    2: ['следопыт'],
    3: ['рейнджер', 'следопыт', 'лекарь', 'лучник', 'мечник', 'боевой маг']
}

def compute_score(hero, task_type, salary=0, weight_quality=1, weight_time=1, weight_politeness=1, salary_weight = 500):
    """
    Функция для вычисления "стоимости" героя для выполнения конкретного поручения с учетом
    типа поручения и роли героя.
    
    Параметры:
    - hero: строка с данными героя (Series).
    - task_type: тип поручения.
    - roles_for_tasks: словарь с ролями для каждого типа поручения.
    - weight_quality, weight_time, weight_politeness: веса для оценки параметров.
    
    Возвращает:
    - Оценку пригодности героя для выполнения поручения.
    """
    
    # Проверим, соответствует ли роль героя типу поручения
    if len(set(hero['Роль'].unique()) & set(roles_for_tasks[task_type])) == 0:
        return 0  # Если роль героя не подходит для этого поручения, возвращаем 0 (герой не подходит)

    # Оценки для конкретного героя и типа поручения
    quality = hero['Оценка за качество'].sum()
    
    time1 = hero['Затрачено часов'].sum()
    time2 = hero['Затрачено дней'].sum()
    speed = hero['Оценка по срокам'].sum()
    politeness = hero['Оценка за вежливость'].sum()
    
    # Вычисляем стоимость героя с учетом всех факторов
    score = (weight_quality * quality) + (speed * weight_time / time1 / time2) + (weight_politeness * politeness)
    return score


# Пример: вычисление для героя
task_type = 0  # Тип поручения
hero = data.loc[(data['Герой'] == 'Агата') & (data['Тип поручения'] == task_type)]


print(compute_score(hero, task_type))


task_type = 0  # Например, тип поручения 0
heroes = data['Герой'].unique()
# Пример: вычисление для героя
for name in heroes:
    hero = data.loc[(data['Герой'] == name) & (data['Тип поручения'] == task_type)]
    for role in hero['Роль'].to_list():
        score = compute_score(hero.loc[hero['Роль'] == role], task_type)
        print(f"Герой: {name}, Роль: {role}, Оценка: {score}")


cases['Нормированная сумма'] = (cases['Сумма вознаграждения'] - cases['Сумма вознаграждения'].min()) / (cases['Сумма вознаграждения'].max() - cases['Сумма вознаграждения'].min())

# Для каждого типа поручения решим задачу на подбор героев
def optimize_team_for_task(task, heroes):
    # Формируем список оценок для каждого героя по данному поручению
    salary = task['Нормированная сумма']
    task_type = task['Тип поручения']
    
    scores = []
    for name in heroes:
        hero = data.loc[(data['Герой'] == name) & (data['Тип поручения'] == task_type)]
        score = []
        for role in hero['Роль'].to_list():
            score.append(compute_score(hero.loc[hero['Роль'] == role], task_type))
        scores.append(max(score))

    # задача сводится к нахождению героев с наибольшими баллами
    # Используем минимизацию, поэтому берём отрицательный score
    c = [-score for score in scores]
    
    # Ограничения на количество героев: 1-4 героя
    A_eq = np.ones((1, len(heroes)))

    if salary < 0.25:
        team_len = 1
    elif salary < 0.5:
        team_len = min(2, len(heroes))
    elif salary < 0.75:
        team_len = min(3, len(heroes))
    else:
        team_len = min(4, len(heroes))

    b_eq = [team_len]  # Не более 4 героев
    x_bounds = [(0, 1) for _ in range(len(heroes))]  # Каждому герою либо 0, либо 1 (выбираем героя или нет)

    res = linprog(c, A_eq=A_eq, b_eq=b_eq, bounds=x_bounds, method='highs')

    # Получаем выбранных героев
    selected_heroes = [heroes[i] for i in range(len(res.x)) if res.x[i] > 0.5]
    #return selected_heroes[max(selected_heroes.keys())]
    return selected_heroes

# Пример
free_heroes = heroes.copy()
team_for_task_0 = optimize_team_for_task(cases.iloc[0], free_heroes)
print("Оптимальная команда для поручения типа:", team_for_task_0)


### **Демонстрация работы на не выполненных заказах**

#### Загрузка данных и определение классов поручений

- Количество членов команды решено формировать в связи с величиной вознаграждения (т.к. принято допущение, что наличие в тексте упоминания "опасности" не отражается на оплате заказа, а также на времени его выполнения) 
- Команды формируются по взвешенной оценке, учитывающей среднее время выполнения заказов, средние оценки
- Герои возвращаются в список доступных в связи по истечении количества дней = max(количество дней на заказ)

cur_cases = pd.read_csv('../data/cases.csv', sep=';')
cur_cases.rename(columns={'\ufeffНомер поручения':'Номер поручения'}, inplace=True)
cur_cases = cur_cases.loc[cur_cases['Выполнено'] == 'нет'].sort_values(by='Дата поручения')

model = joblib.load('../models/classification_model.pkl')

cur_cases['Нормированная сумма'] = (cur_cases['Сумма вознаграждения'] - 5000) / (25000)
cur_cases['Тип поручения'] = model.predict(cur_cases['Описание'])
cur_cases.head()

max_time_interval = cases['Затрачено дней'].max()
teams = []
free_heroes = set(heroes)
query = []
for idx, row in cur_cases.iterrows():
    for date, team in query:
        if (datetime.strptime(row['Дата поручения'], '%Y-%m-%d') - date).days >= max_time_interval:
            free_heroes |= set(team)
    if len(free_heroes) > 0:
        team_for_task = optimize_team_for_task(row, list(free_heroes))
        print(f"Оптимальная команда для поручения {row['Номер поручения']}:", team_for_task)
        query.append((datetime.strptime(row['Дата поручения'], '%Y-%m-%d'), team_for_task))
        teams.append([row['Номер поручения'], team_for_task])
        for name in team_for_task:
            free_heroes -= set(team_for_task)
    else:
        print('Закончились герои')
        break

teams_df = pd.DataFrame(teams, columns=['Номер поручения', 'Предложенная команда'])
teams_df.to_csv('../data/teams.csv', sep=';', index=False)
teams_df
